+++ catalogue/webapp/pages/work.tsx
@@ -1,71 +1,68 @@
-import { NextPage, NextPageContext } from 'next';
-import Router from 'next/router';
+import { GetServerSideProps, NextPage } from 'next';
 import {
   Work as WorkType,
   CatalogueApiError,
 } from '@weco/common/model/catalogue';
-import { workLink } from '@weco/common/services/catalogue/routes';
 import ErrorPage from '@weco/common/views/components/ErrorPage/ErrorPage';
 import Work from '../components/Work/Work';
 import { getWork } from '../services/catalogue/works';
+import GlobalContextProvider, {
+  GlobalContextData,
+  getGlobalContextData,
+} from '@weco/common/views/components/GlobalContextProvider/GlobalContextProvider';
 
 type Props = {
   workResponse: WorkType | CatalogueApiError;
+  globalContextData: GlobalContextData;
 };
 
-export const WorkPage: NextPage<Props> = ({ workResponse }) => {
-  if (
-    workResponse.type === 'Work' ||
-    workResponse.type === 'Collection' ||
-    workResponse.type === 'Section' ||
-    workResponse.type === 'Series'
-  ) {
-    return <Work work={workResponse} />;
-  }
-
-  if (workResponse.type === 'Error') {
-    return (
-      <ErrorPage
-        statusCode={workResponse.httpStatus}
-        title={workResponse.description}
-      />
-    );
-  }
+export const WorkPage: NextPage<Props> = ({
+  workResponse,
+  globalContextData,
+}: Props) => {
+  return (
+    <GlobalContextProvider value={globalContextData}>
+      {workResponse.type === 'Error' && (
+        <ErrorPage
+          statusCode={workResponse.httpStatus}
+          title={workResponse.description}
+        />
+      )}
+      {workResponse.type !== 'Error' && <Work work={workResponse} />}
+    </GlobalContextProvider>
+  );
 };
 
-WorkPage.getInitialProps = async (ctx: NextPageContext) => {
-  const { id } = ctx.query;
+export const getServerSideProps: GetServerSideProps = async context => {
+  const globalContextData = getGlobalContextData(context);
+  const { id } = context.query;
 
   const workResponse = await getWork({
     id,
-    toggles: ctx.query.toggles,
+    toggles: globalContextData.toggles,
   });
 
-  if (workResponse) {
-    if (workResponse.type === 'Redirect') {
-      const { res } = ctx;
-      if (res) {
-        res.writeHead(workResponse.status, {
-          Location: workResponse.redirectToId,
-        });
-        res.end();
-      } else {
-        const link = workLink({ id: workResponse.redirectToId });
-        Router.push(link.href, link.as);
-      }
-    }
-
+  if (workResponse.type === 'NotFound') {
     return {
-      workResponse:
-        workResponse.type === 'Work' ||
-        workResponse.type === 'Error' ||
-        workResponse.type === 'Collection' ||
-        workResponse.type === 'Section' ||
-        workResponse.type === 'Series'
-          ? workResponse
-          : undefined,
+      props: {},
+      notFound: true,
     };
   }
+
+  if (workResponse.type === 'Redirect') {
+    return {
+      props: {},
+      destination: workResponse.redirectToId,
+      permanent: workResponse.status === 301,
+    };
+  }
+
+  return {
+    props: {
+      workResponse,
+      globalContextData,
+    },
+  };
 };
 
 export default WorkPage;
