+++ common/package.json
@@ -18,6 +18,7 @@
     "@next/mdx": "^10.0.1",
     "@popperjs/core": "^2.4.4",
     "@segment/snippet": "^4.10.0",
+    "@types/gtag.js": "^0.0.3",
     "@types/lodash": "^4.14.161",
     "@weco/next-plugin-transpile-modules": "^2.0.1",
     "babel-core": "^7.0.0-bridge.0",
@@ -46,6 +47,7 @@
     "moment-timezone": "^0.5.14",
     "next": "^10.0.1",
     "next-cookies": "^2.0.3",
+    "next-transpile-modules": "^4.1.0",
     "nprogress": "^0.2.0",
     "openseadragon": "^2.4.0",
     "prismic-dom": "^2.1.0",
