+++ common/services/catalogue/swagger.tsx
@@ -1,9 +1,10 @@
-// @flow
 import fetch from 'isomorphic-unfetch';
 
-export default async (useStaging: boolean) =>
+const swagger = async (useStaging: boolean) =>
   fetch(
     `https://api${
       useStaging ? '-stage' : ''
     }.wellcomecollection.org/catalogue/v2/swagger.json`
   ).then(resp => resp.json());
+
+export default swagger;
