+++ common/views/components/PageLayout/PageLayout.js
@@ -2,7 +2,7 @@
 import type { Node } from 'react';
 import type { Url } from '../../../model/url';
 import type { JsonLdObj } from '../JsonLd/JsonLd';
-import { Fragment } from 'react';
+import { Fragment, useContext } from 'react';
 import Head from 'next/head';
 import convertUrlToString from '../../../utils/convert-url-to-string';
 import Header from '../Header/Header';
@@ -11,13 +11,20 @@ import InfoBanner from '../InfoBanner/InfoBanner';
 import CookieNotice from '../CookieNotice/CookieNotice';
 import NewsletterPromo from '../NewsletterPromo/NewsletterPromo';
 import Footer from '../Footer/Footer';
-import GlobalAlertContext from '../GlobalAlertContext/GlobalAlertContext';
 // $FlowFixMe (tsx)
 import PopupDialogContext from '../PopupDialogContext/PopupDialogContext';
 // $FlowFixMe (tsx)
 import PopupDialog from '../PopupDialog/PopupDialog';
 import OpeningTimesContext from '../OpeningTimesContext/OpeningTimesContext';
 import Space from '../styled/Space';
+import GlobalAlertContext from '../GlobalAlertContext/GlobalAlertContext';
+import { museumLd, libraryLd } from '../../../utils/json-ld';
+import { collectionVenueId } from '../../../services/prismic/hardcoded-id';
+import {
+  getParseCollectionVenueById,
+  openingHoursToOpeningHoursSpecification,
+} from '../../../services/prismic/opening-times';
+import { wellcomeCollectionGallery } from '../../../model/organization';
 
 type SiteSection =
   | 'collections'
@@ -64,6 +71,28 @@ const PageLayout = ({
       : 'Wellcome Collection | A free museum and library exploring health and human experience';
 
   const absoluteUrl = `https://wellcomecollection.org${urlString}`;
+  const globalAlert = useContext(GlobalAlertContext);
+  const popupDialog = useContext(PopupDialogContext);
+  const openingTimes = useContext(OpeningTimesContext);
+  const galleries = getParseCollectionVenueById(
+    openingTimes,
+    collectionVenueId.galleries.id
+  );
+  const library = getParseCollectionVenueById(
+    openingTimes,
+    collectionVenueId.libraries.id
+  );
+  const galleriesOpeningHours = galleries && galleries.openingHours;
+  const libraryOpeningHours = library && library.openingHours;
+  const wellcomeCollectionGalleryWithHours = {
+    ...wellcomeCollectionGallery,
+    ...openingHoursToOpeningHoursSpecification(galleriesOpeningHours),
+  };
+  const wellcomeLibraryWithHours = {
+    ...wellcomeCollectionGallery,
+    ...openingHoursToOpeningHoursSpecification(libraryOpeningHours),
+  };
+
   return (
     <Fragment>
       <Head>
@@ -132,6 +161,20 @@ const PageLayout = ({
             type="application/rss+xml"
           />
         )}
+        <script
+          type="application/ld+json"
+          dangerouslySetInnerHTML={{
+            __html: JSON.stringify(
+              museumLd(wellcomeCollectionGalleryWithHours)
+            ),
+          }}
+        />
+        <script
+          type="application/ld+json"
+          dangerouslySetInnerHTML={{
+            __html: JSON.stringify(libraryLd(wellcomeLibraryWithHours)),
+          }}
+        />
       </Head>
 
       <div>
@@ -140,19 +183,16 @@ const PageLayout = ({
           Skip to main content
         </a>
         <Header siteSection={siteSection} />
-        <GlobalAlertContext.Consumer>
-          {globalAlert =>
-            globalAlert &&
-            globalAlert.isShown === 'show' &&
-            (!globalAlert.routeRegex ||
-              urlString.match(new RegExp(globalAlert.routeRegex))) && (
-              <InfoBanner text={globalAlert.text} cookieName="WC_globalAlert" />
-            )
-          }
-        </GlobalAlertContext.Consumer>
-        <PopupDialogContext.Consumer>
-          {({ isShown, ...props }) => isShown && <PopupDialog {...props} />}
-        </PopupDialogContext.Consumer>
+
+        {globalAlert &&
+          globalAlert.isShown === 'show' &&
+          (!globalAlert.routeRegex ||
+            urlString.match(new RegExp(globalAlert.routeRegex))) && (
+            <InfoBanner text={globalAlert.text} cookieName="WC_globalAlert" />
+          )}
+
+        {popupDialog.isShown && <PopupDialog {...popupDialog} />}
+
         <div id="main" className="main" role="main">
           {children}
         </div>
@@ -166,17 +206,14 @@ const PageLayout = ({
             <NewsletterPromo />
           </Space>
         )}
-        <OpeningTimesContext.Consumer>
-          {openingTimes => (
-            <Footer
-              hide={hideFooter}
-              openingTimes={openingTimes}
-              upcomingExceptionalOpeningPeriods={
-                openingTimes.upcomingExceptionalOpeningPeriods
-              }
-            />
-          )}
-        </OpeningTimesContext.Consumer>
+
+        <Footer
+          hide={hideFooter}
+          openingTimes={openingTimes}
+          upcomingExceptionalOpeningPeriods={
+            openingTimes.upcomingExceptionalOpeningPeriods
+          }
+        />
       </div>
     </Fragment>
   );
